Implemented the state controller idle and base movement +cleanup
From the Player State to a Super State to a Sub State they pass the values defined in playerState. The PlayerState functions are called in the player script, and any attached code in each of the states are called whenever you're in one of the states. Some states may even override the normal behavior of the functions, and the state machines manages what state you're in. There's conditionals to change state from one to another in each state, streamlining the logic and making the animation controller super easy and straightforward. The state pattern is easier on the system and easier to code logic without getting tangled in mechanics through conditionals.